refactor: ArticleSection 컴포넌트 분리
